requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
Pillow
seaborn
tqdm
opencv-python
